Move request callbacks to Redis handler
