+++ cvat/apps/engine/utils.py
@@ -21,7 +21,6 @@ import urllib.parse
 from collections import defaultdict, namedtuple
 from collections.abc import Callable, Generator, Iterable, Mapping, Sequence
 from contextlib import nullcontext, suppress
-from enum import StrEnum, auto
 from itertools import islice
 from multiprocessing import cpu_count
 from pathlib import Path
@@ -41,7 +40,7 @@ from redis.lock import Lock
 from rq.job import Job as RQJob
 
 from cvat.apps.engine.types import ExtendedRequest
-from cvat.apps.redis_handler.utils import rq_job_will_be_retried
+from cvat.utils.exceptions import parse_exception_message
 
 if TYPE_CHECKING:
     from _typeshed import StrPath
@@ -50,11 +49,6 @@ if TYPE_CHECKING:
 
 Import = namedtuple("Import", ["module", "name", "alias"])
 log = logging.getLogger(__name__)
-
-
-class RequestStatusEnum(StrEnum):
-    SUCCEEDED = auto()
-    FAILED = auto()
 
 
 def get_request_target_django_model_by_enum(target: "RequestTarget") -> type[Model]:
@@ -66,45 +60,6 @@ def get_request_target_django_model_by_enum(target: "RequestTarget") -> type[Mod
         RequestTarget.JOB: Job,
     }
     return request_target_to_model[target]
-
-
-def send_request_succeeded_signal(
-    rq_job: RQJob,
-    connection: Any,
-    result: Any,
-) -> None:
-    from cvat.apps.engine import signals
-    from cvat.apps.engine.background import BaseResourceExporter
-
-    _ = signals.request_succeeded.send_robust(
-        sender=BaseResourceExporter,
-        request_id=rq_job.id,
-        status=RequestStatusEnum.SUCCEEDED,
-        message=None,
-    )
-
-
-def send_request_failed_signal(
-    rq_job: RQJob,
-    connection: Any,
-    exc_type: type[BaseException],
-    exc_value: BaseException,
-    exc_traceback: Any,
-) -> None:
-    from cvat.apps.engine import signals
-    from cvat.apps.engine.background import BaseResourceExporter
-
-    if rq_job_will_be_retried(rq_job=rq_job):
-        return
-
-    _ = signals.request_failed.send_robust(
-        sender=BaseResourceExporter,
-        request_id=rq_job.id,
-        status=RequestStatusEnum.FAILED,
-        message=parse_exception_message(
-            "".join(traceback.format_exception_only(exc_type, exc_value))
-        ),
-    )
 
 
 def parse_imports(source_code: str):
@@ -208,20 +163,6 @@ def parse_specific_attributes(specific_attributes):
         if parsed_specific_attributes
         else dict()
     )
-
-
-def parse_exception_message(msg: str) -> str:
-    parsed_msg = msg
-    try:
-        if "ErrorDetail" in msg:
-            # msg like: 'rest_framework.exceptions.ValidationError:
-            # [ErrorDetail(string="...", code=\'invalid\')]\n'
-            parsed_msg = msg.split("string=")[1].split(", code=")[0].strip('"')
-        elif msg.startswith("rest_framework.exceptions."):
-            parsed_msg = msg.split(":")[1].strip()
-    except Exception:  # nosec
-        pass
-    return parsed_msg
 
 
 def process_failed_job(rq_job: RQJob) -> str:
