+++ cvat/apps/redis_handler/signals.py
@@ -1,0 +1,8 @@
+# Copyright (C) CVAT.ai Corporation
+#
+# SPDX-License-Identifier: MIT
+
+from django.dispatch import Signal
+
+request_succeeded = Signal()
+request_failed = Signal()
